Displaying daily currencies

+++ Wallet/Sources/Services/Network/Models/OperationApiModel.swift
@@ -10,8 +10,8 @@ public struct OperationApiModel: Codable {
         case income = "INCOME"
         case spending = "SPENDING"
     }
-    public var id: Int64?
-    public var walletId: Int64?
+    public var id: Int?
+    public var walletId: Int?
     public var type: ModelType?
     /** Категория операции */
     public var categoryId: Int64?
