+++ Wallet/Sources/Services/Network/Models/PersonApiModel.swift
@@ -5,7 +5,7 @@
 import Foundation
 
 public struct PersonApiModel: Codable {
-    public var id: Int64?
+    public var id: Int?
     public var email: String?
     public var balance: Decimal?
     public var income: Decimal?
